chore: rename editable-email to email-editor
Refs: #287
