+++ src/components/pages/account-page/current-user-info/editable-email/email-editor/index.tsx
